changed queries to generate a subtable
Whenever the user makes changes the parameters defining valid monsters to consider for value calculation, a subtable is created containing only the applicable monsters. This is being done to make queries to calculate specific values on that subset quicker and more simple.

Note: This has altered how the callback function works. So the query results count is no longer properly set.